+++ src/controls/baseCombo.ts
@@ -444,5 +444,3 @@ export default abstract class WUPBaseComboControl<
     super.gotRemoved();
   }
 }
-
-// todo mouseClick filter only by LeftButton